requirements.txt
aniso8601==9.0.1
click==8.0.4
dataclasses==0.8
Flask==2.0.3
Flask-HTTPAuth==4.7.0
Flask-RESTful==0.3.9
gunicorn==20.1.0
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
PyJWT==2.4.0
pytz==2022.1
six==1.16.0
typing-extensions==4.1.1
Werkzeug==2.0.3
zipp==3.6.0
